Extract the AST-specific parts from the parser into ParserAst
