+++ src/trufflesom/compiler/Symbol.java
@@ -24,7 +24,7 @@
 
 package trufflesom.compiler;
 
-enum Symbol {
+public enum Symbol {
   NONE, Integer, Double, Not, And, Or, Star, Div, Mod, Plus, Minus, Equal, More, Less,
   Comma, At, Per, NewBlock, EndBlock, Colon, Period, Exit, Assign, NewTerm,
   EndTerm, Pound, Primitive, Separator, STString, Identifier, Keyword,
